requirements.txt
certifi==2025.1.31
fastapi==0.115.11
langchain_core
langgraph
openai==1.65.4
protobuf~=4.25.0
pydantic==2.10.6
pymongo==4.6.3
reportlab==4.3.1
tavily_python==0.5.1
uvicorn[standard]==0.34.0
websockets==12.0
google-generativeai==0.8.4